requirements.txt
python-telegram-bot==20.7
opencv-python-headless==4.8.1.78
moviepy==1.0.3
Pillow==10.1.0
python-dotenv==1.0.0
numpy>=1.21.0,<2.0.0
flask==3.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23